reorg test suite
