+++ tests/Tests.FeatureManagement/DotnetFeatureManagementSchema.json
@@ -1,0 +1,90 @@
+﻿{
+  "FeatureManagement": {
+    "OnTestFeature": true,
+    "OffTestFeature": false,
+    "ConditionalFeature": {
+      "EnabledFor": [
+        {
+          "Name": "Test",
+          "Parameters": {
+            "P1": "V1"
+          }
+        }
+      ]
+    },
+    "ContextualFeature": {
+      "EnabledFor": [
+        {
+          "Name": "ContextualTest",
+          "Parameters": {
+            "AllowedAccounts": [
+              "abc"
+            ]
+          }
+        }
+      ]
+    },
+    "AnyFilterFeature": {
+      "RequirementType": "Any",
+      "EnabledFor": [
+        {
+          "Name": "Test",
+          "Parameters": {
+            "Id": "1"
+          }
+        },
+        {
+          "Name": "Test",
+          "Parameters": {
+            "Id": "2"
+          }
+        }
+      ]
+    },
+    "AllFilterFeature": {
+      "RequirementType": "all",
+      "EnabledFor": [
+        {
+          "Name": "Test",
+          "Parameters": {
+            "Id": "1"
+          }
+        },
+        {
+          "Name": "Test",
+          "Parameters": {
+            "Id": "2"
+          }
+        }
+      ]
+    },
+    "FeatureUsesFiltersWithDuplicatedAlias": {
+      "RequirementType": "all",
+      "EnabledFor": [
+        {
+          "Name": "DuplicatedFilterName"
+        },
+        {
+          "Name": "Percentage",
+          "Parameters": {
+            "Value": 100
+          }
+        }
+      ]
+    },
+    "CustomFilterFeature": {
+      "EnabledFor": [
+        {
+          "Name": "CustomTargetingFilter",
+          "Parameters": {
+            "Audience": {
+              "Users": [
+                "Jeff"
+              ]
+            }
+          }
+        }
+      ]
+    }
+  }
+}
