+++ tests/Tests.FeatureManagement/Tests.FeatureManagement.csproj
@@ -49,6 +49,9 @@
     <None Update="MicrosoftFeatureManagement.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
+    <None Update="DotnetFeatureManagementSchema.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
